refactor(body): ordenar imports y aplicar buenas prácticas
