+++ src/Components/Body/TestimonialsMobile.jsx
@@ -1,12 +1,13 @@
-import testimonials from './testimonials.data'
+import testimonials from '../../data/testimonials.data'
 import { useState } from 'react'
-
-/**
- * @returns {import('react').JSX.Element}
- */
 
 const FIRST_INDEX = 0
 const NEXT_INCREMENT = 1
+
+/**
+ * Componente TestimonialsMobile para testimonios en móvil.
+ * @returns {import('react').JSX.Element} Componente TestimonialsMobile.
+ */
 
 const TestimonialsMobile = () => {
   const [current, setCurrent] = useState(FIRST_INDEX)
@@ -26,22 +27,22 @@ const TestimonialsMobile = () => {
   const testimonial = testimonials[current]
 
   return (
-    <div className="max-w-[90%] md:max-w-md mx-auto bg-gray-200 rounded-lg p-6 text-center relative">
+    <div className="max-w-[90%] md:max-w-md mx-auto bg-background-light rounded-lg p-6 text-center relative">
       <img
         src={testimonial.avatar}
         alt={testimonial.name}
         className="w-16 h-16 rounded-full mx-auto mb-4"
       />
-      <p className="text-gray-600 mb-4">"{testimonial.text}"</p>
+      <p className="text-muted mb-4">"{testimonial.text}"</p>
       <h3 className="font-bold text-primary-bg mb-1">{testimonial.name}</h3>
-      <span className="text-gray-400 mb-6 block">{testimonial.role}</span>
+      <span className="text-muted mb-6 block">{testimonial.role}</span>
       <div className="flex justify-center space-x-3 mb-6">
-        {testimonials.map((testimonialItem, index) => (
+        {testimonials.map((testimonialItem, id) => (
           <button
             key={testimonialItem.id}
             type="button"
-            onClick={() => handleSelect(index)}
-            className={`w-3 h-3 rounded-full ${index === current ? 'border-orange-400 bg-orange-400' : 'bg-gray-50'}`}
+            onClick={() => handleSelect(id)}
+            className={`w-3 h-3 rounded-full ${id === current ? 'border-accent bg-accent' : 'bg-background-light'}`}
             aria-label={`Go to testimonial ${testimonialItem.id}`}
           />
         ))}
@@ -49,7 +50,7 @@ const TestimonialsMobile = () => {
       <button
         type="button"
         onClick={handleNext}
-        className="cursor-pointer bg-orange-400 hover:bg-orange-500 text-white px-6 py-3 rounded-full font-semibold transition-colors"
+        className="cursor-pointer bg-accent hover:bg-accent text-light px-6 py-3 rounded-full font-semibold transition-colors"
       >
         Get Started
       </button>
