+++ src/Components/Body/features.data.js
@@ -1,17 +1,20 @@
 export const features = [
   {
+    id: 1,
     number: '01',
     title: 'Track company-wide progress',
     description:
       'See how your day-to-day tasks fit into the wider vision. Go from tracking progress at the milestone level all the way down to the smallest of details. Never lose sight of the bigger picture again.',
   },
   {
+    id: 2,
     number: '02',
     title: 'Advanced built-in reports',
     description:
       'Set internal delivery estimates and track progress toward company goals. Our customisable dashboard helps you build out the reports you need to keep key stakeholders informed.',
   },
   {
+    id: 3,
     number: '03',
     title: 'Everything you need in one place',
     description:
